consolidate IndexManager with IndexService
